Fix command ijection vulnerability
Fix passing of unescaped text to the arguments of notification command
in dcnnt/plugins/notifications.py

This is a high-risk vulnerability that can lead to RCE and allow arbitrary code
execution on the server while the notification showing.

All text received over the network MUST be escaped before being transmitted to
the database or to the shell command arguments.